Added optimization service and entities to store result in database

+++ src/main/java/com/resource/optimization/entity/dto/CreateProjectDto.java
@@ -16,7 +16,7 @@ public class CreateProjectDto {
     private String name;
 
     @NotNull
-    private Double finalCost;
+    private Float finalCost;
 
     private List<CreateTaskDto> tasks;
 }
